requirement.txt
Tensorflow-cpu==2.12.o
Nummpy
Pillow
Streamlit